+++ src/index.js
@@ -1,4 +1,4 @@
 // import whole ECharts package when prebuilding the bundled version
 import 'echarts'
-
-module.exports = './components/ECharts.vue'
+import ECharts from './components/ECharts.vue'
+export default ECharts